requirements-testing.txt
# Testing Dependencies for Azure Web Application Demo
# Install with: pip install -r requirements-testing.txt

# Core testing framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.0.0

# Flask testing utilities
pytest-flask>=1.2.0

# HTTP testing
requests>=2.28.0
responses>=0.23.0

# Azure testing (optional, for integration tests)
azure-mgmt-costmanagement>=4.0.0
azure-identity>=1.12.0
azure-mgmt-resource>=23.0.0

# Security testing
bandit>=1.7.0
safety>=2.3.0

# Code quality
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0

# Coverage reporting
coverage>=7.0.0
